PN-19193: updated resources for private link isolation in pn-core envs
